Fix protected signed backfill staging
