+++ .github/workflows/targeted-signed-reencode.yml
@@ -3041,10 +3041,14 @@ jobs:
                "$RUNNER_TEMP/canonical-refresh-bundle.json" > /dev/null; then
             test -z "$(git -C "$RULESPEC_CHECKOUT" status --porcelain)"
           else
-            "${workflow_python[@]}" \
-              axiom-encode/scripts/prepare_signed_backfill.py stage \
-              "$RULESPEC_CHECKOUT" \
-              --corpus-root axiom-corpus
+            /opt/axiom-verification/axiom-encode-signing-supervisor \
+              --trusted-signing-roots /opt/axiom-verification/signing-trust-roots.json \
+              --trusted-python-runtime-root /opt/axiom-verification/python \
+              --trusted-python-import-root /opt/axiom-verification/python/lib/python3.13/site-packages \
+              --trusted-python-package-root /opt/axiom-verification/python/lib/python3.13/site-packages/axiom_encode \
+              -- /opt/axiom-verification/axiom-encode stage-signed-backfill \
+              --repo "$RULESPEC_CHECKOUT" \
+              --corpus-path "$GITHUB_WORKSPACE/axiom-corpus"
             git -C "$RULESPEC_CHECKOUT" \
               -c core.hooksPath=/dev/null \
               -c user.name=axiom-encode-bot \
