+++ scripts/provision_verification_supervisor.py
@@ -477,7 +477,7 @@ def _resolve_trusted_git(raw_git: Path) -> Path:
 def _install_trusted_git_wrapper(
     tool_directory: Path, interpreter: Path, trusted_git: Path
 ) -> Path:
-    """Install a read-only Git command broker into the isolated PATH."""
+    """Install a narrowly scoped Git command broker into the isolated PATH."""
 
     wrapper = tool_directory / "git"
     content = (
@@ -494,7 +494,8 @@ def _install_trusted_git_wrapper(
         "        raise SystemExit('trusted git wrapper requires a command after -C')\n"
         "    cursor = 2\n"
         "allowed = {\n"
-        "    'cat-file', 'diff', 'grep', 'log', 'ls-files', 'ls-tree', 'merge-base',\n"
+        "    'add', 'cat-file', 'diff', 'grep', 'log', 'ls-files', 'ls-tree',\n"
+        "    'merge-base',\n"
         "    'remote', 'rev-list', 'rev-parse', 'show', 'status',\n"
         "}\n"
         "if cursor >= len(args) or args[cursor] not in allowed:\n"
@@ -517,6 +518,18 @@ def _install_trusted_git_wrapper(
         "        and all(part not in {'', '.', '..'} for part in value.split('/'))\n"
         "    )\n"
         "def allowed_options():\n"
+        "    if command == 'add':\n"
+        "        repository = args[1] if args[:1] == ['-C'] else ''\n"
+        "        return (\n"
+        "            cursor == 2\n"
+        "            and os.path.isabs(repository)\n"
+        "            and os.path.normpath(repository) == repository\n"
+        "            and os.path.realpath(repository) == repository\n"
+        "            and os.path.isdir(repository)\n"
+        "            and len(options) >= 2\n"
+        "            and options[0] == '--'\n"
+        "            and all(safe_path(value) for value in options[1:])\n"
+        "        )\n"
         "    if command == 'cat-file':\n"
         "        return options == ['--batch'] or (\n"
         "            len(options) == 2\n"
@@ -524,6 +537,10 @@ def _install_trusted_git_wrapper(
         "            and re.fullmatch(r'[0-9a-f]{40}', options[1]) is not None\n"
         "        )\n"
         "    if command == 'diff':\n"
+        "        if options == [\n"
+        "            '--cached', '--name-only', '--no-renames', '-z',\n"
+        "        ]:\n"
+        "            return True\n"
         "        if (\n"
         "            options[:3] == ['--binary', 'HEAD', '--']\n"
         "            and all(safe_path(value) for value in options[3:])\n"
@@ -544,21 +561,21 @@ def _install_trusted_git_wrapper(
         "    if command == 'grep':\n"
         "        if options[:4] != ['-z', '-l', '-a', '-F']:\n"
         "            return False\n"
-        "        cursor = 4\n"
+        "        option_cursor = 4\n"
         "        patterns = 0\n"
         "        while (\n"
-        "            cursor + 1 < len(options)\n"
-        "            and options[cursor] == '-e'\n"
-        "            and safe_path(options[cursor + 1])\n"
+        "            option_cursor + 1 < len(options)\n"
+        "            and options[option_cursor] == '-e'\n"
+        "            and safe_path(options[option_cursor + 1])\n"
         "            and patterns < 64\n"
         "        ):\n"
-        "            cursor += 2\n"
+        "            option_cursor += 2\n"
         "            patterns += 1\n"
         "        return (\n"
         "            patterns > 0\n"
-        "            and len(options) == cursor + 2\n"
-        "            and safe_ref(options[cursor])\n"
-        "            and options[cursor + 1] == '--'\n"
+        "            and len(options) == option_cursor + 2\n"
+        "            and safe_ref(options[option_cursor])\n"
+        "            and options[option_cursor + 1] == '--'\n"
         "        )\n"
         "    if command == 'log':\n"
         "        return options == [\n"
@@ -669,6 +686,8 @@ def _install_trusted_git_wrapper(
         "    '-c', 'submodule.recurse=false',\n"
         "]\n"
         "def worktree_sensitive():\n"
+        "    if command == 'add':\n"
+        "        return True\n"
         "    if command == 'status':\n"
         "        return True\n"
         "    if command != 'diff':\n"
