+++ src/axiom_encode/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.2.1647"
+__version__ = "0.2.1648"
 # Axiom Encode - AI-assisted RuleSpec encoding
 # Self-contained encoding infrastructure -- no external plugin dependencies.
